Vuelvo a crear modelos con base limpia. Arregla issues con algunos id que no eran BIGINT Al editar reporte se recalcula el valor isAsleep de las mediciones (x si edita período de sueño) Además se recalculan los totales por lo anterior El piechart de valores sobre el límite muestra 0% cuando no hay medidas (en vez de NeuN) Deshabilita usuario Pregunta antes de salir de la aplicación.

+++ trunk/UDA-HTA/DataAccess/PatientDataAccess.cs
@@ -21,8 +21,8 @@ namespace DataAccess
             {
                 using (var udaContext = new patient_info_dbEntities())
                 {
-                    ObjectParameter lastIdPatient = new ObjectParameter("id", typeof(int));
-                    ObjectParameter lastIdDevRef = new ObjectParameter("id", typeof(int));
+                    ObjectParameter lastIdPatient = new ObjectParameter("id", typeof(long));
+                    ObjectParameter lastIdDevRef = new ObjectParameter("id", typeof(long));
 
                     var sex = SexType.M.ToString();
                     if (p.Sex != null && ((p.Sex.Value == SexType.F) || (p.Sex.Value == SexType.M)))
