+++ trunk/UDA-HTA/Entities/InvestigationSearch.cs
@@ -4,7 +4,7 @@ namespace Entities
 {
     public class InvestigationSearch
     {
-        public int IdInvestigation { get; set; }
+        public long IdInvestigation { get; set; }
         public string Name { get; set; }
         public DateTime CreationDate { get; set; }
         public string Comment { get; set; }
@@ -13,7 +13,7 @@ namespace Entities
         {
         }
 
-        public InvestigationSearch(int id, string name, DateTime date, string comment)
+        public InvestigationSearch(long id, string name, DateTime date, string comment)
         {
             IdInvestigation = id;
             Name = name;
